+++ Changelog.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 # 0.2.0 [Unreleased]
 The beta release of Neanderthal.
 - [x] Prerendered markdown syntax highlighting.
-- [ ] Markdown Nunjucks tag- towards the unified file format.
+- [x] Markdown Nunjucks tag- towards the unified file format.
 - [x] Error handling in Nunjucks templates.
 - [ ] Improvements to Frontmatter processing
   - [ ] Internal type safety